Z averaging of the many zonal flow instance
It reads a lat-radius non-uniform data of many zonal flow snapshots from pickle files, performs an interpolation onto to a regular uniform grid, averages the zonal flow in both north and south hemispheres and plots the data as a function of time.  It can be used to analyze torsional oscillations.