Function for truncating degree using SHTNS lib
Can be used to truncate (low-pass) the harmonic degree of lat-lon data set. It uses the SHTNS library. The script assumes the library is properly installed in the computer and can simply be imported. For using in a script, do ===> "import sys
sys.path.append('path to function'); from Truncate_degree_SHTNS_func import shtns_truncate"